requirements.txt
asgiref==3.9.1
crispy-bootstrap4==2025.6
Django==5.2.4
django-crispy-forms==2.4
django-environ==0.12.0
django-widget-tweaks==1.5.0
pillow==11.3.0
PyMySQL==1.1.1
sqlparse==0.5.3
typing_extensions==4.14.1
